requirements/pip_requirements.txt
absl-py==2.3.0
accelerate==1.7.0
adabelief-pytorch==0.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmasher==1.9.2
colorama==0.4.6
colorcet==3.1.0
colorspacious==1.1.2
configmypy==0.2.0
conflictfree==0.1.8
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
diffusers==0.33.1
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
GIFt @ git+https://github.com/qiauil/GIFt@bf93a005d1e5854778d2a34a36f40226d1dd31c8
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-bolts==0.7.0
lightning-utilities==0.14.3
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
moviepy==2.2.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
neuraloperator==0.3.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pbdl @ git+https://github.com/tum-pbs/pbdl-dataset@3fcc5cdabb3ef2052455548c883849c6c23b072f
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
POT==0.9.5
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorly==0.9.0
tensorly-torch==0.5.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
torch-fidelity==0.3.0
torch_harmonics==0.7.4
torchfsm @ git+https://github.com/qiauil/torchfsm.git@2e996775e356e21945c2a13467a9c922055ea8ef
torchmetrics==1.7.1
torchprofile==0.0.4
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vape4d==0.2.7
vedo==2025.5.3
vtk==9.4.2
wandb==0.19.11
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.22.0
